+++ adapters/IronSource/Public/Headers/IronSourceAdapter.h
@@ -1,10 +1,16 @@
+// Copyright 2017 Google Inc.
 //
-//  IronSourceAdapter.h
-//  IronSourceAdMobAdapter
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//  Created by Itai Asaf on 26/03/2017.
-//  Copyright © 2017 ironSource Ltd. All rights reserved.
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 #ifndef IronSourceAdapter_h
 #define IronSourceAdapter_h