+++ adapters/Maio/MaioAdapter/GADMMaioError.m
@@ -11,35 +11,38 @@
 @implementation GADMMaioError
 
 + (NSError *)errorWithDescription:(NSString *)description {
-    description = [description copy];
-    NSDictionary *userInfo = @{NSLocalizedDescriptionKey : description, NSLocalizedFailureReasonErrorKey : description};
-    NSError *error = [NSError errorWithDomain:GADMMaioErrorDomain code:0 userInfo:userInfo];
-    return error;
+  description = [description copy];
+  NSDictionary *userInfo = @{
+    NSLocalizedDescriptionKey : description,
+    NSLocalizedFailureReasonErrorKey : description
+  };
+  NSError *error =
+      [NSError errorWithDomain:GADMMaioErrorDomain code:0 userInfo:userInfo];
+  return error;
 }
-
 
 /**
  *  MaioFailReason の文字列表記を取得します。
  */
 + (NSString *)stringFromFailReason:(MaioFailReason)failReason {
-    switch (failReason) {
-        case MaioFailReasonUnknown:
-            return @"Unknown";
-        case MaioFailReasonNetworkConnection:
-            return @"NetworkConnection";
-        case MaioFailReasonNetworkServer:
-            return @"NetworkServer";
-        case MaioFailReasonNetworkClient:
-            return @"NetworkClient";
-        case MaioFailReasonSdk:
-            return @"Sdk";
-        case MaioFailReasonDownloadCancelled:
-            return @"DownloadCancelled";
-        case MaioFailReasonAdStockOut:
-            return @"AdStockOut";
-        case MaioFailReasonVideoPlayback:
-            return @"VideoPlayback";
-    }
+  switch (failReason) {
+  case MaioFailReasonUnknown:
+    return @"Unknown";
+  case MaioFailReasonNetworkConnection:
+    return @"NetworkConnection";
+  case MaioFailReasonNetworkServer:
+    return @"NetworkServer";
+  case MaioFailReasonNetworkClient:
+    return @"NetworkClient";
+  case MaioFailReasonSdk:
+    return @"Sdk";
+  case MaioFailReasonDownloadCancelled:
+    return @"DownloadCancelled";
+  case MaioFailReasonAdStockOut:
+    return @"AdStockOut";
+  case MaioFailReasonVideoPlayback:
+    return @"VideoPlayback";
+  }
 }
 
 @end