+++ adapters/AppLovin/CHANGELOG.md
@@ -1,10 +1,16 @@
 # AppLovin Adapter for Google Mobile Ads SDK for iOS
 
+## Version 4.6.1.0
+- Verified compatibility with AppLovin SDK 4.6.1.
+
+## Version 4.6.0.0
+- Verified compatibility with AppLovin SDK 4.6.0.
+
 ## Version 4.5.1.0
 - Verified compatibility with AppLovin SDK 4.5.1.
 
 ## Version 4.4.1.1
-- Add support for banner ads.
+- Added support for banner ads.
 
 ## Version 4.4.1.0
 - Verified compatibility with AppLovin SDK 4.4.1.