+++ adapters/Facebook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Facebook Audience Network Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Version 4.27.1.0
+- Verified compatibility with FAN SDK 4.27.1.
+
+## Version 4.27.0.0
+- Verified compatibility with FAN SDK 4.27.0.
+
 ## Version 4.26.1.0
 - Verified compatibility with FAN SDK 4.26.1.