+++ adapters/IronSource/CHANGELOG.md
@@ -1,4 +1,6 @@
-# IronSource Mediation Adapter for Google Mobile Ads SDK for iOS
+# ironSource Adapter for Google Mobile Ads SDK for iOS
 
-## Version 6.7.4
-- Initial release. Supports reward-based video ads and interstitial ads.
+## x.y.z.0
+
+- First release!
+- Added support for banners, interstitials, etc.