+++ adapters/Maio/CHANGELOG.md
@@ -1,6 +1,5 @@
 # maio Adapter for Google Mobile Ads SDK for iOS
 
-## x.y.z.0
-
+## Version 1.2.18.0
 - First release!
-- Adds support for banners, interstitials, etc.
+- Added support for interstitials and rewarded videos ads.